apps/backend-core/pyproject.toml
[tool.poetry]
name = "spine"
version = "0.1.0"
description = "VTE Logic Spine - The Proof Authority"
authors = ["VTE Team <dev@vte.local>"]
package-mode = true

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = "^0.30.0"
pydantic = "^2.6.0"
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pytest = "^8.3.4"
httpx = "^0.26.0"
stripe = "^8.1.0"
PyJWT = "^2.8.0"
email-validator = "^2.1.0"
python-multipart = "^0.0.20"
openai = "^1.10.0"
jinja2 = "^3.1.0"

msgraph-core = "^1.0.0"
google-api-python-client = "^2.116.0"
boto3 = "^1.34.0"
sentry-sdk = "^1.40.0"
redis = "^5.0.1"
celery = "^5.3.6"
pandas = "^2.2.0"
pypdf = "^4.2.0"
openpyxl = "^3.1.2"
pydantic-settings = "^2.0.0"
slowapi = "^0.1.9"
twilio = "^8.0.0"
sendgrid = "^6.11.0"
google-auth = "^2.0.0"
pyotp = "^2.9.0"
zxcvbn = "^4.5.0"
python-dotenv = "^1.2.1"
bcrypt = "4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
